# refactor: no longer use activeTrackId, activeScopeTrackId