--- a/src/unstable/asyncComputed.ts
+++ b/src/unstable/asyncComputed.ts
@@ -1,5 +1,4 @@
 import { Computed } from '../computed.js';
-import { nextTrackId } from '../effect.js';
 import { Dependency, endTrack, link, shallowPropagate, startTrack, SubscriberFlags } from '../system.js';
 import { asyncCheckDirty } from './asyncSystem.js';
 
@@ -37,21 +36,13 @@ export class AsyncComputed<T = any> extends Computed {
 	async update(): Promise<boolean> {
 		try {
 			startTrack(this);
-			const trackId = nextTrackId();
 			const oldValue = this.currentValue;
 			const generator = this.getter(oldValue);
 			let current = await generator.next();
 			while (!current.done) {
 				const dep = current.value;
-				if (dep.lastTrackedId !== trackId) {
-					dep.lastTrackedId = trackId;
-					link(dep, this);
-				}
+				link(dep, this);
 				current = await generator.next();
-
-				// if (this.flags & SubscriberFlags.Recursed) {
-				// 	return await this.get() !== oldValue;
-				// }
 			}
 			const newValue = await current.value;
 			if (oldValue !== newValue) {