`activeTrackId` and `activeScopeTrackId` were originally designed to quickly skip unnecessary dynamic dependency updates. However, since dynamic dependencies are low-frequency events, using `isValidLink()` with a time complexity of O(n) instead and completely removing the `activeTrackId` and `activeScopeTrackId` logic actually results in generally much faster.

### Benchmark Results

#### main branch

Memory Usage: 2498.22 KB

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | `  1.40 µs/iter` | `  1.40 µs` | `  1.40 µs` | `  1.41 µs` | `  1.41 µs` |
| propagate: 1 * 10    | `  7.05 µs/iter` | `  7.04 µs` | `  7.05 µs` | `  7.05 µs` | `  7.05 µs` |
| propagate: 1 * 100   | ` 61.90 µs/iter` | ` 61.87 µs` | ` 61.91 µs` | ` 61.92 µs` | ` 61.97 µs` |
| propagate: 10 * 1    | ` 12.54 µs/iter` | ` 12.53 µs` | ` 12.55 µs` | ` 12.55 µs` | ` 12.55 µs` |
| propagate: 10 * 10   | ` 68.97 µs/iter` | ` 68.29 µs` | ` 68.92 µs` | ` 72.33 µs` | ` 91.08 µs` |
| propagate: 10 * 100  | `617.76 µs/iter` | `614.33 µs` | `618.88 µs` | `637.96 µs` | `652.67 µs` |
| propagate: 100 * 1   | `123.08 µs/iter` | `121.92 µs` | `122.79 µs` | `135.42 µs` | `148.25 µs` |
| propagate: 100 * 10  | `682.61 µs/iter` | `679.21 µs` | `684.00 µs` | `693.00 µs` | `724.08 µs` |
| propagate: 100 * 100 | `  6.16 ms/iter` | `  6.15 ms` | `  6.16 ms` | `  6.21 ms` | `  6.21 ms` |

#### This PR

Memory Usage: 2413.55 KB

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | `  1.28 µs/iter` | `  1.27 µs` | `  1.28 µs` | `  1.29 µs` | `  1.30 µs` |
| propagate: 1 * 10    | `  6.25 µs/iter` | `  6.24 µs` | `  6.25 µs` | `  6.25 µs` | `  6.26 µs` |
| propagate: 1 * 100   | ` 54.46 µs/iter` | ` 54.39 µs` | ` 54.49 µs` | ` 54.57 µs` | ` 54.59 µs` |
| propagate: 10 * 1    | ` 11.24 µs/iter` | ` 11.24 µs` | ` 11.24 µs` | ` 11.24 µs` | ` 11.25 µs` |
| propagate: 10 * 10   | ` 60.63 µs/iter` | ` 60.60 µs` | ` 60.64 µs` | ` 60.66 µs` | ` 60.69 µs` |
| propagate: 10 * 100  | `546.65 µs/iter` | `543.54 µs` | `548.00 µs` | `561.58 µs` | `570.88 µs` |
| propagate: 100 * 1   | `109.85 µs/iter` | `108.58 µs` | `109.67 µs` | `118.63 µs` | `152.00 µs` |
| propagate: 100 * 10  | `605.66 µs/iter` | `602.33 µs` | `607.04 µs` | `618.83 µs` | `629.33 µs` |
| propagate: 100 * 100 | `  5.44 ms/iter` | `  5.43 ms` | `  5.44 ms` | `  5.47 ms` | `  5.47 ms` |